Add: back end web socket to front when game satart
