+++ front/src/components/base/AppContainer.vue
@@ -14,7 +14,6 @@
 import SidebarVue from "../sidebar/Sidebar.vue";
 import Header from "../header/Header.vue";
 import MainContainer from "./MainContainer.vue";
-import { socketHandler } from "@/services/userInfo";
 import { onMounted } from "vue";
 
 export default {
@@ -23,23 +22,6 @@ export default {
     SidebarVue,
     Header,
     MainContainer,
-  },
-  setup() {
-    function connectWebSocket() {
-      const token = localStorage.getItem("token");
-      if (token) {
-        const socketInstance = new socketHandler(token);
-        socketInstance.socket.connect();
-        socketInstance.socket.emit("login", null);
-      }
-    }
-    onMounted(() => {
-      connectWebSocket();
-    });
-
-    return {
-      connectWebSocket,
-    };
   },
 };
 </script>
